Workaround for `moveToFailed()` failing on clusters (#1)
I tried to check for `instanceof redis.Cluster`, but for some reason it's not evaluating to true, so I had to go with an options check.  It's difficult to debug as well, because we're not running in cluster mode locally, only on AWS.  See inline comments for more detail - if anyone has any ideas on a better check, let me know.